moved methods called in post_setup_multi_commodity to SLC base class
