+++ h2integrate/control/control_strategies/system_level/demand_following_control.py
@@ -128,15 +128,6 @@ class DemandFollowingControl(SystemLevelControlBase):
         self.converter_upstreams = converter_upstreams
         self.converter_tech_names = converter_tech_names
 
-    def dict_values_to_flat_list(self, dictionary):
-        flat_list = []
-        for v in dictionary.values():
-            if isinstance(v, set):
-                v = list(v)
-
-            flat_list.extend(v)
-        return flat_list
-
     def get_setpoints_for_commodity_subset(
         self, inputs, outputs, commodity, commodity_demand, tech_subset: list | set | None = None
     ):
@@ -295,136 +286,6 @@ class DemandFollowingControl(SystemLevelControlBase):
             )
             return result
         return tech_groups_demand
-
-    def _check_demand_tech_group_connections(self, converter_tech_names, missing_input_techs):
-        successful = True
-        commodities_for_missing_techs = {
-            tech: self._get_commodity_for_tech(tech) for tech in list(missing_input_techs)
-        }
-        commodities_out = self.dict_values_to_flat_list(commodities_for_missing_techs)
-
-        if self.commodity not in list(commodities_out):
-            warnings.warn(
-                "none of the demand commodities are made by missing techs",
-                UserWarning,
-                stacklevel=3,
-            )
-            successful = False
-
-        missing_tech_downstreams = [
-            list(nx.descendants(self.technology_graph, tech)) for tech in missing_input_techs
-        ]
-        missing_tech_downstreams_shared = {self.demand_tech}
-        other_converters = converter_tech_names - missing_input_techs
-
-        for downstream in missing_tech_downstreams:
-            missing_tech_downstreams_shared = missing_tech_downstreams_shared & set(downstream)
-            # TODO: check that no other converters are inbetween
-            if set(downstream) & other_converters:
-                warnings.warn(
-                    "theres an extra converter between the missing techs and the demand",
-                    UserWarning,
-                    stacklevel=3,
-                )
-                successful = False
-        if not missing_tech_downstreams_shared or len(missing_tech_downstreams_shared) > 1:
-            warnings.warn("something unexpected happened", UserWarning, stacklevel=3)
-            successful = False
-
-        missing_converter_tech = missing_input_techs & converter_tech_names
-        if len(missing_converter_tech) > 1:
-            warnings.warn(
-                "unsure how code will work with multiple converters connected to demand",
-                UserWarning,
-                stacklevel=3,
-            )
-            successful = False
-        if not missing_converter_tech:
-            warnings.warn("should have a converter before demand ...", UserWarning, stacklevel=3)
-            successful = False
-
-        for m0 in list(missing_converter_tech):
-            for m1 in list(missing_input_techs - missing_converter_tech):
-                m0_upstream = nx.has_path(self.technology_graph, m0, m1)
-                m1_upstream = nx.has_path(self.technology_graph, m1, m0)
-                if not m0_upstream or m1_upstream:
-                    warnings.warn("these technologies arent connected", UserWarning, stacklevel=3)
-                    successful = False
-        return successful
-
-    def _find_demand_tech_group(self, converters, converter_upstreams):
-        found_input_techs = self.dict_values_to_flat_list(converter_upstreams)
-        missing_input_techs = set(self.input_techs) - set(found_input_techs)
-
-        converter_tech_names = {v[1] for v in list(converters)}
-
-        successful = self._check_demand_tech_group_connections(
-            converter_tech_names, missing_input_techs
-        )
-        if not successful:
-            msg = "A bug may exist. Please refer to earlier warnings"
-            warnings.warn(msg, UserWarning, stacklevel=3)
-
-        input_comps = {self.demand_tech} - missing_input_techs
-        missing_techs_to_downstreams = {
-            tech: list(nx.descendants(self.technology_graph, tech)) for tech in missing_input_techs
-        }
-        missing_tech_downstreams = self.dict_values_to_flat_list(missing_techs_to_downstreams)
-        non_input_components = set(missing_tech_downstreams) - input_comps
-
-        unique_number = int(len(converter_upstreams) + 1)
-        group_name = f"{self.commodity}-{int(unique_number)}"
-
-        missing_converter_tech = missing_input_techs & converter_tech_names
-
-        # all techs in group, include non-controllable ones (like combiners)
-        all_techs_in_group = list(non_input_components | missing_input_techs)
-        # input techs in group except for the converter
-        non_converter_input_techs_in_group = list(missing_input_techs - missing_converter_tech)
-        demand_group = {group_name: all_techs_in_group}
-
-        return non_converter_input_techs_in_group, demand_group
-
-    def _find_group_for_non_input_techs(self, grouped_techs):
-        # Get the nodes of the technology graph that aren't a controllable technology
-        def get_group_for_tech(tech_name):
-            group = [grp for grp, techs in grouped_techs.items() if tech_name in techs]
-            if len(group) == 0:
-                msg = f"Cannot find simplified group for technology {tech_name}"
-                raise ValueError(msg)
-            return group[0]
-
-        techs_to_groups = {}
-        conversion_factor_keys = []
-
-        non_input_techs = (
-            set(self.technology_graph.nodes) - set(self.input_techs) - {self.demand_tech}
-        )
-        # Also add these technologies to the reversed_group_techs
-
-        for non_t in list(non_input_techs):
-            up_techs = set(self.technology_graph.predecessors(non_t)) - non_input_techs
-            down_techs = set(self.technology_graph.successors(non_t)) - non_input_techs
-
-            commod = None
-            if up_techs:
-                for t in list(up_techs):
-                    commod = self.technology_graph.edges[t, non_t].get("commodity", None)
-                    if commod is not None:
-                        # Add these technologies to the reversed_group_techs
-                        techs_to_groups[non_t] = get_group_for_tech(t)
-                        break
-
-            if down_techs and commod is None:
-                for t in list(down_techs):
-                    commod = self.technology_graph.edges[non_t, t].get("commodity", None)
-                    if commod is not None:
-                        # Add these technologies to the reversed_group_techs
-                        techs_to_groups[non_t] = get_group_for_tech(t)
-                        break
-            # Add conversion factors of 1 for the technologies that are non_input_techs
-            conversion_factor_keys.append((commod, non_t, commod))
-        return conversion_factor_keys, techs_to_groups
 
     def compute(self, inputs, outputs):
         if not self.multi_commodity_system:
